divide run_frag into run_pw, run_lcao, and run_linearscaling
