+++ ABACUS.develop/source/run_linearscaling.h
@@ -1,28 +1,20 @@
 //==========================================================
 // AUTHOR : mohan
-// DATE : 2008-11-07
+// DATE : 2021-02-01
 //==========================================================
-#ifndef RUN_FRAG_H
-#define RUN_FRAG_H
+#ifndef RUN_LINEARSCALING_H
+#define RUN_LINEARSCALING_H
 
 #include "src_pw/tools.h"
 #include "input.h"
 
-class Run_Frag
+class Run_linearscaling
 {
 
 public:
 
-    Run_Frag();
-    ~Run_Frag();
-
-    static void init(void);
-	
-	// perform Linear Combination of Atomic Orbitals (LCAO) calculations
-    void LCAO_line(void);
-
-	// perform plane wave basis calculations
-    void plane_wave_line(void);
+    Run_linearscaling();
+    ~Run_linearscaling();
 
 	// perform linear-scaling calculations for sparse matrix H and S
 	// useless now
